# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service for CRUD",
        "title": "CRUD MySql API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9093",
    "basePath": "/",
    "paths": {
        "/user/deleteUser/{id}": {
            "delete": {
                "description": "Delete a user by id was input",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete User"
                ],
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Delete user by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        },
        "/user/findAll": {
            "get": {
                "description": "Get all user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get All User"
                ],
                "summary": "Get all user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        },
        "/user/findOne": {
            "get": {
                "description": "Tes url with query param",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test Param"
                ],
                "summary": "Tes url with query param",
                "parameters": [
                    {
                        "type": "string",
                        "description": "First Name",
                        "name": "firstName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Last Name",
                        "name": "lastName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/findOne/{id}": {
            "get": {
                "description": "Get user by id was input",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get User"
                ],
                "summary": "Get a user by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Get user by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        },
        "/user/saveUser": {
            "post": {
                "description": "Create a new user with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Save User"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "Create user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        },
        "/user/updateUser/{id}": {
            "put": {
                "description": "Update a user with the id and input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update User"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "description": "Update user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "Update user",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.ResponseUser"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "user.ResponseUser": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.User"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}